Bug fix de tabelas Tela, Evento e TipoEvento.

+++ src/main/java/br/com/laranja/springcrud/infrastructure/repository/ProjetoRepository.java
@@ -3,5 +3,5 @@ package br.com.laranja.springcrud.infrastructure.repository;
 import br.com.laranja.springcrud.domain.model.Projeto;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProjetoRepository extends CrudRepository<Projeto, Integer> {
+public interface ProjetoRepository extends CrudRepository<Projeto, Long> {
 }
